Refactored directory entry into its own reusable file
